backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
ultralytics
numpy
pydantic
cloudinary
python-dotenv
